% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_book.R
\name{build_books}
\alias{build_books}
\title{Build books.}
\usage{
build_books(manuals_folder = "manuals", manuals, all_manuals, verbose = TRUE)
}
\arguments{
\item{manuals_folder}{Defaults to \code{manuals}}

\item{manuals}{Vector of manuals to build, e.g. \code{"R-intro.texi"}}

\item{all_manuals}{Vector of all manuals, e.g. \code{c("R-intro.texi", "R-data.texi")}}
}
\description{
Build books.
}
