% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{build_main_website}
\alias{build_main_website}
\title{Build main website with all books.}
\usage{
build_main_website(manuals_folder = "manuals")
}
\arguments{
\item{manuals_folder}{Defaults to "manuals"}
}
\value{
Path to rendered site
}
\description{
Build main website with all books.
}
