% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_html_to_md}
\alias{convert_html_to_md}
\title{Convert intermediate HTML to markdown, using a lua filter}
\usage{
convert_html_to_md(input_file, verbose = FALSE)
}
\arguments{
\item{input_file}{Name of input file}

\item{verbose}{If TRUE, prints progress messages}
}
\value{
Invisible null - writes into a set of HTML files
}
\description{
Convert intermediate HTML to markdown, using a lua filter
}
