% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_to_md}
\alias{convert_to_md}
\title{Convert all html files in a directory to markdown}
\usage{
convert_to_md(path = "temp", verbose = FALSE)
}
\arguments{
\item{path}{Directory containing the HTML files}

\item{verbose}{If TRUE, prints progress messages}
}
\value{
Invisible NULL - writes a set of markdown files
}
\description{
Convert all html files in a directory to markdown
}
