% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_manuals.R
\name{download_manuals}
\alias{download_manuals}
\title{Download manuals}
\usage{
download_manuals(manual = "R-exts.texi", destdir = "data", r_ref = NULL)
}
\arguments{
\item{manual}{Filename of manual to download. This should be the \code{.texi} file, e.g. \code{R-exts.texi}}

\item{destdir}{Destination directory}

\item{r_ref}{Ref in the wch/r-source mirror to download from. \code{NULL} (the
default) resolves to the latest released version of R via
\code{\link[=latest_r_release_ref]{latest_r_release_ref()}}. Use \code{"trunk"} for R-devel, a release branch like
\code{"R-4-6-branch"}, or an exact tag like \code{"tags/R-4-6-1"}.}
}
\value{
downloads the file
}
\description{
Download manuals
}
