% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_info.R
\name{make_info}
\alias{make_info}
\title{Make manual from \code{.texi} file using \code{makeinfo}}
\usage{
make_info(manual, input_dir = "data", output_dir = "temp", verbose = FALSE)
}
\arguments{
\item{manual}{Name of manual, e.g. \code{R-exts.texi}}

\item{input_dir}{Input directory}

\item{output_dir}{Destination directory}

\item{verbose}{If TRUE, prints progress messages}
}
\value{
Runs \code{makeinfo} and generates html files
}
\description{
Make manual from \code{.texi} file using \code{makeinfo}
}
