% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_manuals.R
\name{pre_process_manuals}
\alias{pre_process_manuals}
\title{Pre-process the texi files.}
\usage{
pre_process_manuals(destdir = "data")
}
\arguments{
\item{destdir}{Destination directory}
}
\description{
At the moment the only pre-processing is to remove \verb{@group} statements.
In PDF format this ensures line breaks don't occur in the middle of code blocks,
but in HTML this causes multiple \code{pre} statements.
}
