% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_book.R
\name{update_quarto_yaml}
\alias{update_quarto_yaml}
\title{Build a book}
\usage{
update_quarto_yaml(x, index, all_manuals, verbose = FALSE)
}
\arguments{
\item{x}{manual name}

\item{index}{index file}

\item{all_manuals}{Used to make the navbar, i.e. cross-referencing all the other books}
}
\value{
A modified yaml object
}
\description{
Build a book
}
\keyword{Internal}
