% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_manuals.R
\name{download_manuals}
\alias{download_manuals}
\title{Download manuals}
\usage{
download_manuals(
  manual = "R-exts.texi",
  destdir = "data",
  base_url = "https://raw.githubusercontent.com/wch/r-source/trunk"
)
}
\arguments{
\item{manual}{Filename of manual to download. This should be the \code{.texi} file, e.g. \code{R-exts.texi}}

\item{destdir}{Destination directory}

\item{base_url}{Base URL}
}
\value{
downloads the file
}
\description{
Download manuals
}
