% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_manual.R
\name{process_manual}
\alias{process_manual}
\title{Process manual source file to prepare book to build.}
\usage{
process_manual(
  manual = "R-exts.texi",
  folder = tolower(sub(".texi", "", manual)),
  manuals_folder = "manuals",
  .quicktest = FALSE,
  .make_info = !.quicktest,
  .download = !.quicktest,
  verbose = TRUE
)
}
\arguments{
\item{manual}{Name of the manual to download and process (a \code{.texi}) file}

\item{folder}{sub-folder for intermediaries files per manual. Default to name
of manual}

\item{manuals_folder}{folder where all the intermediary folders are stored.
\code{manuals_folder} is the parent \code{folder}.}

\item{.quicktest}{If TRUE, removes some of the HTML files after \code{makeinfo}
step, for faster quarto rendering on partial site}

\item{.make_info}{If TRUE, convert \code{.texi} to \code{.html}}

\item{.download}{If TRUE, download the manual files}

\item{verbose}{If TRUE, prints progress messages}
}
\value{
NULL.
}
\description{
Downloads, pre-processes and converts the source \code{.texi} files into the necessary
structure and files are there to build the book with Quarto
}
